Corrections + prise en charge des tokens pour accèder à l'API

+++ src/main/java/com/esgi/controllers/ContactController.java
@@ -6,6 +6,7 @@ package com.esgi.controllers;
 
 import com.esgi.model.Contact;
 import com.esgi.services.ContactService;
+import com.esgi.services.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
@@ -16,13 +17,24 @@ public class ContactController {
     @Autowired
     private ContactService contactService;
 
+    @Autowired
+    private UserService userService;
+
     @RequestMapping(value = "/add", method = RequestMethod.POST)
-    public void addContact(@RequestBody Contact contact) {
-        contactService.addContact(contact);
+    public void addContact(@RequestParam String token, @RequestBody Contact contact) {
+        Long iduser = userService.getIdFromToken(token);
+        if (iduser != null) {
+            contact.setIduser(iduser);
+            contact.setAccepted(false);
+            contactService.addContact(contact);
+        }
     }
 
     @RequestMapping(value = "/remove", method = RequestMethod.POST)
-    public void removeContact(@RequestParam Long iduser, @RequestParam Long idcontact) {
-        contactService.removeContact(iduser, idcontact);
+    public void removeContact(@RequestParam String token, @RequestParam Long idcontact) {
+        Long iduser = userService.getIdFromToken(token);
+        if (iduser != null) {
+            contactService.removeContact(iduser, idcontact);
+        }
     }
 }
